Move HudCommsWidget to a different dir
